Make GroupContext local to the group JSON presenter
