+++ h/traversal/__init__.py
@@ -63,12 +63,7 @@ shouldn't return model objects directly).
 from h.traversal.annotation import AnnotationContext, AnnotationRoot
 from h.traversal.auth_client import AuthClientRoot
 from h.traversal.bulk_api import BulkAPIRoot
-from h.traversal.group import (
-    GroupContext,
-    GroupRoot,
-    GroupUpsertContext,
-    GroupUpsertRoot,
-)
+from h.traversal.group import GroupRoot, GroupUpsertContext, GroupUpsertRoot
 from h.traversal.organization import OrganizationContext, OrganizationRoot
 from h.traversal.profile import ProfileRoot
 from h.traversal.root import Root
@@ -80,7 +75,6 @@ __all__ = (
     "AnnotationRoot",
     "AuthClientRoot",
     "BulkAPIRoot",
-    "GroupContext",
     "GroupRoot",
     "GroupUpsertContext",
     "GroupUpsertRoot",
